refactor: 重构 scan_select 实现，提取共享工具并增强功能
- 新建 ScanSelectUtils 工具类，统一目录扫描和 glob 匹配逻辑
- 移除 TaskOptionGenerator 和 MaaProcessor 中的重复代码
- 增强 ProcessScanSelectPipeline 支持数组内对象的 attach 替换
- 简化 MaaInterface.cs 中的 pipeline 处理逻辑
- 避免不必要的 JObject 与 Dictionary 转换，提升性能
