+++ MFAAvalonia/Helper/TaskOptionGenerator.cs
@@ -1099,144 +1099,7 @@ public class TaskOptionGenerator(TaskQueueViewModel viewModel, Action saveConfig
     /// </summary>
     private void ScanDirectory(MaaInterface.MaaInterfaceOption interfaceOption)
     {
-        if (string.IsNullOrEmpty(interfaceOption.ScanDir) || string.IsNullOrEmpty(interfaceOption.ScanFilter))
-            return;
-
-        try
-        {
-            var basePath = MaaProcessor.ResourceBase;
-            var scanPath = interfaceOption.ScanDir.Replace("{PROJECT_DIR}", basePath);
-            scanPath = Path.GetFullPath(scanPath);
-
-            if (!Directory.Exists(scanPath))
-            {
-                LoggerHelper.Warning($"Scan directory does not exist: {scanPath}");
-                interfaceOption.Cases ??= new List<MaaInterface.MaaInterfaceOptionCase>();
-                return;
-            }
-
-            var files = EnumerateFilesWithGlob(scanPath, interfaceOption.ScanFilter);
-            var newCases = new List<MaaInterface.MaaInterfaceOptionCase>();
-
-            foreach (var file in files.OrderBy(f => f))
-            {
-                var fileName = Path.GetFileName(file);
-                var caseItem = new MaaInterface.MaaInterfaceOptionCase
-                {
-                    Name = fileName,
-                    Label = fileName
-                };
-                newCases.Add(caseItem);
-            }
-
-            interfaceOption.Cases = newCases;
-
-            // 初始化所有 case 的显示名称
-            foreach (var caseItem in interfaceOption.Cases)
-            {
-                caseItem.InitializeDisplayName();
-            }
-        }
-        catch (Exception ex)
-        {
-            LoggerHelper.Error($"Failed to scan directory: {ex.Message}");
-            interfaceOption.Cases ??= new List<MaaInterface.MaaInterfaceOptionCase>();
-        }
-    }
-
-    /// <summary>
-    /// 使用 glob 模式枚举文件
-    /// </summary>
-    private List<string> EnumerateFilesWithGlob(string basePath, string pattern)
-    {
-        var results = new List<string>();
-
-        // 处理 ** 递归模式
-        if (pattern.Contains("**/"))
-        {
-            var parts = pattern.Split(new[] { "**/" }, StringSplitOptions.None);
-            if (parts.Length == 2)
-            {
-                var baseDir = parts[0];
-                var filePattern = parts[1];
-
-                var searchPath = string.IsNullOrEmpty(baseDir)
-                    ? basePath
-                    : Path.Combine(basePath, baseDir);
-
-                if (Directory.Exists(searchPath))
-                {
-                    // 递归搜索所有子目录
-                    var allFiles = Directory.EnumerateFiles(searchPath, "*", SearchOption.AllDirectories);
-                    foreach (var file in allFiles)
-                    {
-                        var fileName = Path.GetFileName(file);
-                        if (MatchesGlobPattern(fileName, filePattern))
-                        {
-                            results.Add(file);
-                        }
-                    }
-                }
-            }
-        }
-        else if (pattern.Contains("*"))
-        {
-            // 处理包含通配符的目录路径，如 "settings/*.json"
-            var dirPart = Path.GetDirectoryName(pattern) ?? "";
-            var filePart = Path.GetFileName(pattern);
-
-            if (string.IsNullOrEmpty(dirPart))
-            {
-                // 当前目录
-                if (Directory.Exists(basePath))
-                {
-                    results.AddRange(Directory.EnumerateFiles(basePath, filePart));
-                }
-            }
-            else
-            {
-                // 子目录
-                var searchPath = Path.Combine(basePath, dirPart);
-                if (Directory.Exists(searchPath))
-                {
-                    results.AddRange(Directory.EnumerateFiles(searchPath, filePart));
-                }
-            }
-        }
-        else
-        {
-            // 简单模式，直接搜索
-            if (Directory.Exists(basePath))
-            {
-                results.AddRange(Directory.EnumerateFiles(basePath, pattern));
-            }
-        }
-
-        return results;
-    }
-
-    /// <summary>
-    /// 检查文件名是否匹配 glob 模式
-    /// </summary>
-    private bool MatchesGlobPattern(string fileName, string pattern)
-    {
-        // 将 glob 模式转换为正则表达式
-        var regexPattern = pattern
-            .Replace(".", "\\.")
-            .Replace("*", ".*")
-            .Replace("?", ".");
-
-        regexPattern = "^" + regexPattern + "$";
-
-        try
-        {
-            return Regex.IsMatch(fileName, regexPattern, RegexOptions.IgnoreCase);
-        }
-        catch
-        {
-            // 如果正则表达式无效，使用简单的字符串比较
-            return fileName.Equals(pattern, StringComparison.OrdinalIgnoreCase);
-        }
+        ScanSelectUtils.ScanDirectory(interfaceOption, MaaProcessor.ResourceBase);
     }
 
     // Helper Creation Methods
